Separate the tests
